+++ R/mock_forest.R
@@ -73,7 +73,7 @@ mock_app_forest_mm <- function() {
       numeric_numeric_functions = numeric_numeric_functions,
       numeric_factor_functions = numeric_factor_functions,
       visit_var = "VISIT",
-      value_var = c("VALUE1", "VALUE2"),
+      value_vars = c("VALUE1", "VALUE2"),
       subjid_var = "SUBJID",
       cat_var = "PARCAT"
     )